feat: adjusted dependencies and package name
