+++ app/src/main/kotlin/br/com/snr/financeappsnr/MainApplication.kt
