+++ app/src/main/kotlin/br/com/snr/financeappsnr/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package br.com.snr.financeappsnr
+package br.com.snr.financeappsnr.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,23 +6,26 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import br.com.snr.financeappsnr.ui.theme.FinanceAppSnrTheme
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.rememberNavController
+import br.com.snr.financeappsnr.presentation.navigation.NavigationGraph
+import br.com.snr.financeappsnr.presentation.ui.theme.FinanceAppSnrTheme
 
 class MainActivity : ComponentActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
+            val navController = rememberNavController()
+
             FinanceAppSnrTheme {
-                // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Greeting("Android")
+                    CallNavigation(navController = navController)
                 }
             }
         }
@@ -30,17 +33,6 @@ class MainActivity : ComponentActivity() {
 }
 
 @Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    FinanceAppSnrTheme {
-        Greeting("Android")
-    }
+private fun CallNavigation(navController: NavHostController) {
+    NavigationGraph(navController = navController)
 }
