+++ app/src/main/kotlin/br/com/snr/financeappsnr/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package br.com.snr.financeappsnr.ui.theme
+package br.com.snr.financeappsnr.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
