+++ app/src/remote/kotlin/br/com/snr/financeappsnr/di/DataModule.kt
@@ -1,0 +1,9 @@
+package br.com.snr.financeappsnr.di
+
+import org.koin.dsl.module
+
+object DataModule {
+    val module = module {
+        //TODO: adicionar dependencia da implementação do repositorio aqui
+    }
+}
